is-tiffany-keypoints-detector/requirements.txt
is-wire
is-msgs
numpy
opencv-python-headless
ultralytics
opencensus-ext-zipkin